requirements.txt
Flask==2.0.1
pyodbc==4.0.32
Flask-Caching==1.10.1
python-dotenv==0.19.0
